requirements.txt
pandas==2.1.2
selenium==4.28.1
openpyxl==3.1.5
python-dotenv